Shift hotkey logic from MainWindow class to GUI class
