+++ src/seedu/addressbook/ui/Gui.java
@@ -1,5 +1,6 @@
 package seedu.addressbook.ui;
 
+import javafx.application.Platform;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
@@ -8,6 +9,12 @@ import seedu.addressbook.Main;
 
 import java.io.File;
 import java.io.IOException;
+
+import javax.swing.KeyStroke;
+
+import com.tulskiy.keymaster.common.HotKey;
+import com.tulskiy.keymaster.common.HotKeyListener;
+import com.tulskiy.keymaster.common.Provider;
 
 /**
  * The GUI of the App
@@ -23,6 +30,8 @@ public class Gui {
 
     private MainWindow mainWindow;
     private String version;
+    
+    private Provider hotkeyManager;
 
     public Gui(Logic logic, String version) {
         this.logic = logic;
@@ -30,8 +39,15 @@ public class Gui {
     }
 
     public void start(Stage stage, Stoppable mainApp) throws IOException {
+        initializeSystemHotkey(stage);
         mainWindow = createMainWindow(stage, mainApp);
         mainWindow.displayWelcomeMessage(version, logic.getStorageFilePath());
+    }
+    
+    public void stop(){
+        /* Unbind hotkey */
+        hotkeyManager.reset();
+        hotkeyManager.stop();
     }
 
     private MainWindow createMainWindow(Stage stage, Stoppable mainApp) throws IOException{
@@ -50,5 +66,18 @@ public class Gui {
         mainWindow.setMainApp(mainApp);
         return mainWindow;
     }
-
+    
+    /**
+     * Initialize HotkeyManager and bind show/hide hotkey
+     */
+    private void initializeSystemHotkey(Stage stage) {
+        hotkeyManager = Provider.getCurrentProvider(false);
+        hotkeyManager.register(KeyStroke.getKeyStroke("control alt D"), new HotKeyListener() {
+            public void onHotKey(HotKey hotKey) {
+                System.out.println(hotKey);
+                Platform.runLater(()-> {stage.setIconified(true);});
+            }
+        });   
+    }
+    
 }
